Move cross contract return value tests to own example
